factor wrapper into classes (#3)
Factor the code wrapping functionality into a set of classes that match the
specifications that the library takes as input. Also add support for rubocop
to police the quality of the code in the library, and make several
adjustments according to its recommendations. Finally, add support for
bundler installation and dependency management.