add version tests (#11)
Add tests for versioning to ensure full coverage. Due to the loading order
of modules, the inclusion of the version module in the gemspec must be
removed so that the module can be included after the coverage tools. An extra
test has been added to ensure that the gemspec and module versions do not
diverge.