add class-level include (#19)
While specifying include lists on a per-element basis results in a more
efficient inclusion list in some cases, it can be tedious if everything in
the wrapped library is packaged in a single include. This change adds the
option to provide a class-level include list that removes the need for an
include list for every element.