allow string include list (#21)
In many cases the include list for an element is a single file. In cases
like this, it is more convenient to simply specify a string for the include
list rather than define a list of one element. This change gives this option
to users and normalizes the list on the back end.