add version checks to specs (#29)
Adds an optional version field to scope, class, function, and constant
specifications. This allows the library to detect when a specification
intended for a newer version is given and raise an error if so.

In the future this field might also be used to convert older specs to
newer specifications.