refactor constructor and destructors (#31)
Constructors and destructors were being generated using a separate set of
code, which overlapped with the standard function wrapping code in many
cases. This change refactors the code generation for these special
functions to use the standard wrapping functions, simplifying the logic
flow and removing duplicate code.

Constructors generated automatically based on struct members are still
handled via their own set of functions. These should be refactored to use
the standard functionality as well in the future, but this will require
extra functionality for generating wrappers that is not currently
implemented.