support struct overloading (#39)
Some structs in C libraries may need to be wrapped by different classes
based on some attribute within them. This is supported by adding a list of
rules in the equivalent struct for each derived class such that it can be
routed to the correct wrapping class at runtime. This allows the wrapping
language to use features like runtime polymorphism in a clean way.