remove duplicate constructor definitions (#40)
When the default pointer construction functionality was added, the old code
to create the pointer constructors was not removed. This led to multiple
definitions of a default constructor based on a pointer to the wrapped
struct, which cases compilation issues. The duplicate code has been
removed, and a test added to catch regressions.