drop ruby 2.3 support (#75)
Several gems in use (rubocop and simplecov, for example) require Ruby 2.4
for their newest versions. Dropping support for Ruby 2.3 allows use of
features in the latest versions, for example new rubocop cops.

simplecov has been kept below 0.18 due to a compatibility issue with the
truffleruby 20.0.0 implementation of Float#floor. Once truffleruby 20.1.0
is available on Travis CI, the constraint will be removed. See the relevant
issue https://github.com/oracle/truffleruby/issues/1899 for details.

A test has also been added to ensure that the version date in the ChangeLog
and the date of the gemspec are in sync.