fix function pointer return errors (#80)
Definitions of functions that return a function themselves were incorrectly
generated, dumping the contents of the TypeSpec hash directly into the
code. This is corrected by adding a new function (TypeSpec#cast_expression)
and using it in the function definition generation.

This fixes #76, which can be referenced for more details.