fix actions with no params (#85)
If a `throw-exception` action was defined with a constructor that did not
have any parameters, then an error was raised when the `take` method was
called on the ActionSpec instance. This has been fixed by using
`WrappedFunctionSpec::normalize_spec_hash` to normalize the spec.

This fixes #78, which can be referenced for more details.