fix return value usage in constructors (#86)
Constructors with an `error-check` specified which used the keyword
`return-value` referenced a local variable named `return_val`, but this
variable was not assigned to prior to the usage. This resulted in
segmentation faults upon execution. This has been corrected by using the
equivalent struct member of the class in constructors, and not generating
the local variable at all.

This fixes #84, which can be referenced for more details.