bring coverage back up (#91)
A multi-line conditional responsible for determining whether a parent
class's constructor could be use as the initializer for a child class
was not completely covered, resulting in a drop in coverage when codecov
began to detect this. This change refactors the conditional in question,
and adds a test to ensure that it is covered.