fix return value variables (#93)
Return values were not properly declared, captured, or returned for some
functions resulting in improper functionality in some cases and code that
would not compile in others. This has been rectified so that return values
are properly captured in cases where an error check needs to happen before
returning, even if the check itself does not use the return value.

Some dependencies are also updated in this change to correct functionality
with continuous integration tools.