add version check in executable (#94)
When generating classes using the `bin/wrapture` script, the versions of
the provided specs was not checked. This meant that old versions of
wrapture would still attempt to generate wrappers for specs using features
for newer versions. This check has been added to the executable script so
that an error is thrown when a newer version spec than what the running
version can handle is provided.