start python support (#103)
This change is the beginning of adding Python wrapping support for C code,
in addition to the existing C++ capability. This brings with it a major
re-architecture, which affects the C++ implementation heavily as well.
This is not fully functional Python support, but does bring several other
updates and improvements to the architecture with it.