factor names into separate words
Name handling was previously inconsistent, with some names being assumed to
be in UpperCamelCase, some allowed to be nil or empty, and others with no
convention at all. This change updates the Named module to depend on a name
decomposed into its separate words (as name_words) instead of a string. This
allows names to be used with more flexibility and consistency.

A portability issue in the CToPython module is also fixed in this change.